backend_django/requirements.txt
django==5.1.7
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.16.0
